fix: 消除64个编译警告（宏重复定义加#ifndef守卫、64位类型转换加intptr_t中间层、strncpy缓冲区溢出修复、tomato_gui指针类型修正、续行符尾部空格清理）

+++ AIO_Firmware_PIO/src/app/pc_resource/pc_resource.cpp
@@ -268,7 +268,7 @@ static void pc_resource_message_handle(const char *from, const char *to,
     case APP_MESSAGE_WIFI_CONN:
     {
         Serial.println(F("----->pc_resource_event_notification"));
-        int event_id = (int)message;
+        int event_id = (int)(intptr_t)message;
         switch (event_id)
         {
         case UPDATE_RS_DATA:
